requirements.txt
# requirements.txt

# Core Web Framework & ASGI
fastapi>=0.115.12
uvicorn>=0.34.0
starlette>=0.46.1
anyio>=4.9.0
h11>=0.14.0
websockets>=14.2
sniffio>=1.3.1

# HTTP Clients & Networking
aiohttp>=3.11.16
aiohappyeyeballs>=2.6.1
aiosignal>=1.3.2
attrs>=25.3.0
frozenlist>=1.6.0
multidict>=6.4.3
yarl>=1.20.0
httpx>=0.28.1
httpcore>=1.0.8
h2>=4.2.0
hpack>=4.1.0
hyperframe>=6.1.0
idna>=3.10
certifi>=2025.1.31
dnspython>=2.7.0
propcache>=0.3.1

# Database, ORM & Migrations
SQLAlchemy>=2.0.40
GeoAlchemy2>=0.17.1
alembic>=1.15.2
psycopg>=3.3.2
psycopg-binary>=3.3.2
greenlet>=3.1.1
Mako>=1.3.9
MarkupSafe>=3.0.2
packaging>=24.2

# Supabase Client & Realtime Services
supabase>=2.15.0
postgrest>=1.0.1
gotrue>=2.12.0
realtime>=2.4.2
storage3>=0.11.3
supafunc>=0.9.4

# Authentication, Security & Cryptography
python-jose>=3.4.0
PyJWT>=2.10.1
passlib>=1.7.4
rsa>=4.9
ecdsa>=0.19.1
pyasn1>=0.4.8
email_validator>=2.2.0

# Data Validation & Configuration
pydantic>=2.11.0
pydantic_core>=2.33.0
pydantic-settings>=2.8.1
annotated-types>=0.7.0
typing_extensions>=4.13.0
typing-inspection>=0.4.0
python-dotenv>=1.1.0
python-dateutil>=2.9.0.post0
tzdata>=2025.2
six>=1.17.0
StrEnum>=0.4.15

# Background Tasks, Queues & Caching
celery>=5.5.1
redis>=5.2.1
amqp>=5.3.1
billiard>=4.2.1
kombu>=5.5.3
vine>=5.1.0

# File Uploads, Storage & Media
python-multipart>=0.0.20
python-magic-bin>=0.4.14
pillow>=12.0.0

# Logging, Observability & Debugging
python-json-logger>=3.3.0
structlog>=25.2.0
colorama>=0.4.6
deprecation>=2.1.0

# CLI Utilities & Developer Tools
click>=8.1.8
click-didyoumean>=0.3.1
click-plugins>=1.1.1
click-repl>=0.3.0
prompt_toolkit>=3.0.51
wcwidth>=0.2.13

# Testing Frameworks
pytest>=8.3.5
pytest-mock>=3.14.0
pluggy>=1.5.0
iniconfig>=2.1.0

# Miscellaneous Utilities
uuid>=1.30
